+++ NEWS.md
@@ -11,6 +11,7 @@
 * Add "Maturing" as a lifecycle badge
 * Add new constructor functions to (re)establish object classes for defol and outbreak. This is particularly useful when batch processing with {purrr} functions like `map`
 * Changed the `outbreak` object code to `obr`. And updated several functions to account for the change.
+* Bug fix in `get_defol_events()` to pass non-defoliated series through to final table.
 
 
 # dfoliatR v0.0.3